new_entry.php: fix URL error.
The current code assumes a date was added to the new_entry URL string and
when a user is added, it's always added with a & separator. This results in an
URL error when there is no date.
The proper thing to do is to check if the URL string contains a ? and if not,
consider there is no URL query yet and use a ? separator.