Improve event addition control.
In group mode, every user can add an event into someone else's calendar.
The way to do it:
1) Log-in as user1
2) Display user2 calendar
2) Add an event
3) When displaying this event, the entry "Created by" shows: "user2 (by user1).
Access control is not enabled.

The expected behaviour is to forbid event addition in any calendar except
the owner's one and the ones for which he/she is assistant/admin. Of course,
inviting someone as a participant to an event is not concerned.