Move auth settings to a configuration file.
Until now, authentication parameter settings are spread among
includes/user*.php files, that also contains code.
As a consequence, these are not very well suitable to be moved to some
configuration directory (i.e.: in a system-wide, packaged installation).
This commit moves all settable authentication parameters to a new
file includes/auth-settings.php. In a packaged system-wide installation, this
file can then be moved to a configuration directory and linked from its
original place.
In addition, this commit fixes a bad link in the user-imap.php module and
unifies the authentication function prototypes.