Add files via upload
I had the same problem and could solve the issue for MySQL/MariaDB the following way:

1. Transform all encodings in the database to utf-8 via phpMyAdmin (operating on MySQL and MariaDB).
2. Ensure that the database connection is using utf-8, by modifiying the file includes/dbi4php.php 
   from line 117 onwards:
   add the following lines:
   
   $original_character_set = $c->character_set_name();
	if (strpos(strtolower($original_character_set),"utf") == false)
	{
		$fgcharsetName =  "utf8";
		$resultOfCharsetChange =  $c->set_charset ($fgcharsetName);
	}
	
Side-Effect:
A crude change of the encoding for the tables causes the existing special characters (for instance German Umlaute) to be displayed incorrectly. Everything new will be displayed correctly.
The way I coped with that is to export the database to a file. Change the encoding to utf-8 correct anything (search/replace) that does not look the way it should do and reimport the file into the existing database. That way all data is displayed correctly. Once everything is moved to utf-8 there is very little need to use another character set and whatever language the user sets in the browser, the event data is displayed correctly. In my case I use English and German language browsers and in both cases the special characters are displayed correctly (Umlaute).