Add files via upload
The changes here are small but adress the issue of not being able to export events via 
<webcalendar base path>/publish.php/<user name>.ics
With the changes here (and the conversion to utf-8) the WebCalendar can be integrated into the calendar module of the current Thunderbird release (52.6) , even with German special characters.
The idea was to make sure that a conversion to utf-8 of the data to be sent to thunderbird takes place only if there is a need for conversion to utf8.
Sadly there is still an issue with events that are repeated: They are not shown correctly in thunderbirds calendar module lightning. 
I am working on fixing that issue.