FG: line 453ff: Error message because $minute isset but is not numeric
Error because of PHP 7.x
Just a simple change: The code did check whether $hour was set and assumed that $minute is treated alike, but in PHP 7.x $minute is set and empty but not numeric. To make sure you can multiply its value, you have to do a second check and see if it is numeric.