Creating event causes log, which breaks redirect
When creating a new event, and activate the mailing for yourself ("Event that I create => YES"), then a log is created when $this->SMTPDebug = 2.
This breaks the redirect of the HTML page, and gives the user "confusing" feedback.  Everything works, but the user receives a weird email log, and a tiny link to click to get back to the expected page.

Disabling error logging fixes this.